+++ src/lib/api.ts
@@ -1,0 +1,5 @@
+
+export { CleanOptions } from './tasks/clean';
+export { GitError } from './errors/git-error';
+export { GitResponseError } from './errors/git-response-error';
+export { TaskConfigurationError } from './errors/task-configuration-error';